Extract API interactions into `HttpClient`
