+++ src/utils/index.ts
@@ -1,7 +1,3 @@
-export { authenticateWithCode } from "./authenticate-with-code";
-export { authenticateWithRefreshToken } from "./authenticate-with-refresh-token";
-export { getAuthorizationUrl } from "./get-authorization-url";
-export { getLogoutUrl } from "./get-logout-url";
 export { isRedirectCallback } from "./is-redirect-callback";
 export { memoryStorage } from "./memory-storage";
 export { createPkceChallenge } from "./pkce";
